% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basecoat-package.R
\docType{package}
\name{basecoat-package}
\alias{basecoat}
\alias{basecoat-package}
\title{basecoat: 'Basecoat UI' Components for 'htmltools' and 'shiny'}
\description{
Build web pages and 'shiny' apps from 'Basecoat UI' components. 'Basecoat UI' is a 'Tailwind CSS' version of 'shadcn/ui' that needs no 'React'. Every function returns an 'htmltools' tag, so the components work in static pages, server-rendered apps, and 'shiny'. The stylesheets and scripts ship with the package, so pages work offline.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ricochet-rs.github.io/basecoat/}
  \item \url{https://basecoatui.com}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Josiah Parry \email{josiah.parry@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Ronan Berder (Author of the bundled basecoat-css library) [contributor, copyright holder]
}

}
\keyword{internal}
