% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combobox.R
\name{bc_combobox_group}
\alias{bc_combobox_group}
\title{Combobox Group}
\usage{
bc_combobox_group(title, ...)
}
\arguments{
\item{title}{String. The group title.}

\item{...}{Options from \code{\link[=bc_combobox_option]{bc_combobox_option()}}.}
}
\value{
An htmltools tag.
}
\description{
Create a titled group of combobox options.
}
\examples{
bc_combobox_group(
  "Americas",
  bc_combobox_option("New York"),
  bc_combobox_option("Chicago")
)
}
