% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbd.R
\name{bc_kbd}
\alias{bc_kbd}
\title{Keyboard Key}
\usage{
bc_kbd(label, ...)
}
\arguments{
\item{label}{String. Text on the key, such as \code{"⌘K"}.}

\item{...}{Extra tag attributes.}
}
\value{
An htmltools tag.
}
\description{
Create a keyboard key label.
}
\examples{
bc_kbd("⌘K")

bc_kbd("⏎", `data-icon` = "inline-end")
}
