% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{bc_label}
\alias{bc_label}
\title{Label}
\usage{
bc_label(...)
}
\arguments{
\item{...}{Tag attributes and content. Set \code{`for`} to the ID of the
control the label names.}
}
\value{
An htmltools tag.
}
\description{
Create a label for a form control.
}
\details{
\code{\link[=bc_checkbox]{bc_checkbox()}} and \code{\link[=bc_switch]{bc_switch()}} already include a label.
}
\examples{
bc_label("Your email address", `for` = "email")

bc_label(
  htmltools::tags$input(type = "checkbox", class = "input"),
  "Accept terms and conditions"
)
}
\seealso{
\code{\link[=bc_checkbox]{bc_checkbox()}}, \code{\link[=bc_switch]{bc_switch()}}, \code{\link[=bc_field]{bc_field()}}
}
