% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{bc_pagination}
\alias{bc_pagination}
\alias{bc_pagination_item}
\alias{bc_pagination_ellipsis}
\alias{bc_pagination_previous}
\alias{bc_pagination_next}
\title{Pagination}
\usage{
bc_pagination(..., label = "pagination")

bc_pagination_item(item, href = "#", current = FALSE, ...)

bc_pagination_ellipsis()

bc_pagination_previous(label = "Previous", href = "#")

bc_pagination_next(label = "Next", href = "#")
}
\arguments{
\item{...}{For \code{bc_pagination()}, items from \code{bc_pagination_item()},
\code{bc_pagination_previous()}, \code{bc_pagination_next()} and
\code{bc_pagination_ellipsis()}. For \code{bc_pagination_item()}, extra tag
attributes for the link.}

\item{label}{String. For \code{bc_pagination()}, the accessible name, which
defaults to \code{"pagination"}. For \code{bc_pagination_previous()} and
\code{bc_pagination_next()}, the link text.}

\item{item}{String. The page label.}

\item{href}{String. The link URL. Defaults to \code{"#"}.}

\item{current}{Bool. Whether this is the current page, shown with the
outline style. Defaults to \code{FALSE}.}
}
\value{
An htmltools tag.
}
\description{
Create a row of page links.
}
\examples{
bc_pagination(
  bc_pagination_previous(),
  bc_pagination_item("1", href = "#"),
  bc_pagination_item("2", href = "#", current = TRUE),
  bc_pagination_item("3", href = "#"),
  bc_pagination_ellipsis(),
  bc_pagination_next()
)
}
