% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{bc_styles}
\alias{bc_styles}
\title{Style Packs}
\format{
A character vector of 8 names.
}
\usage{
bc_styles
}
\description{
The bundled style packs.
}
\examples{
bc_styles

lapply(bc_styles, function(style) bc_deps(style, js = FALSE))
}
