% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-switcher.R
\name{bc_theme_switcher}
\alias{bc_theme_switcher}
\title{Theme Switcher}
\usage{
bc_theme_switcher(...)
}
\arguments{
\item{...}{Extra tag attributes.}
}
\value{
An htmltools tag.
}
\description{
Create a button that switches the page between light and dark mode.
}
\examples{
bc_theme_switcher()
}
