% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basecoat-package.R
\docType{package}
\name{basecoat-package}
\alias{basecoat}
\alias{basecoat-package}
\title{basecoat: Basecoat UI Components for htmltools}
\description{
Builds Basecoat UI markup with htmltools tags, and attaches the Basecoat stylesheet and scripts from a CDN as an HTML dependency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://basecoatui.com}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah@ricochet.rs}

Authors:
\itemize{
  \item Josiah Parry \email{josiah@ricochet.rs}
}

}
\keyword{internal}
