% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basecoat-package.R
\docType{package}
\name{basecoat-package}
\alias{basecoat}
\alias{basecoat-package}
\title{basecoat: Beautiful User Interfaces Via 'Basecoat UI'}
\description{
Create beautiful user interfaces through Basecoat UI componenets. Basecoat is a pure-tailwind implementation of shadcn/ui without any 'React'. The package is built on 'htmltools' and designed to work with all UI frameworks with a custom 'shiny' integration to boot. 'CSS' and 'JavaScript' dependencies are included in the package and no 'cdn' is used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ricochet-rs.github.io/basecoat/}
  \item \url{https://basecoatui.com}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@github.com}

Authors:
\itemize{
  \item Josiah Parry \email{josiah.parry@github.com}
}

Other contributors:
\itemize{
  \item Ronan Berder (Author of the bundled basecoat-css library) [contributor, copyright holder]
}

}
\keyword{internal}
