% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basecoat-package.R
\docType{package}
\name{basecoat-package}
\alias{basecoat}
\alias{basecoat-package}
\title{basecoat: Basecoat UI Components}
\description{
Builds Basecoat UI markup with htmltools tags, and attaches the bundled Basecoat stylesheet and scripts as an HTML dependency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://basecoatui.com}
  \item \url{https://ricochet-rs.github.io/basecoat/}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah@ricochet.rs}

Authors:
\itemize{
  \item Josiah Parry \email{josiah@ricochet.rs}
}

Other contributors:
\itemize{
  \item Ronan Berder (Author of the bundled basecoat-css library) [contributor, copyright holder]
}

}
\keyword{internal}
