% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combobox.R
\name{bc_combobox_group}
\alias{bc_combobox_group}
\title{Combobox group}
\usage{
bc_combobox_group(title, ...)
}
\arguments{
\item{title}{String. The group heading.}

\item{...}{Option tags to group under the heading.}
}
\value{
A \verb{<div>} tag of class \verb{option group}.
}
\description{
A headed set of options separated from the rest of the list.
}
\examples{
bc_combobox_group(
  "Americas",
  bc_combobox_option("New York"),
  bc_combobox_option("Chicago")
)
}
