% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combobox.R
\name{bc_combobox_separator}
\alias{bc_combobox_separator}
\title{Combobox separator}
\usage{
bc_combobox_separator()
}
\value{
An \verb{<hr>} tag.
}
\description{
A horizontal rule between combobox groups.
}
\examples{
bc_combobox_separator()
}
