% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{bc_select_option}
\alias{bc_select_option}
\title{Select Option}
\usage{
bc_select_option(value, label = NULL)
}
\arguments{
\item{value}{String. The value of the option.}

\item{label}{String. The display label for the option.}
}
\value{
A list containing option properties.
}
\description{
A helper function to create individual select options.
}
