% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{bc_tab}
\alias{bc_tab}
\title{Tab Panel}
\usage{
bc_tab(..., id = NULL, label, disabled = FALSE)
}
\arguments{
\item{...}{Content for the tab panel.}

\item{id}{String. The ID for this tab panel (optional).}

\item{label}{String. The label for this tab.}

\item{disabled}{Bool. Whether the tab is disabled.}
}
\value{
A list containing tab properties.
}
\description{
A helper function to create individual tab content panels.
}
\examples{
bc_tab("Account details", label = "Account")

bc_tab(
  htmltools::p("Nothing to bill yet."),
  label = "Billing",
  disabled = TRUE
)
}
